REFACTOR: enum 정리
