+++ backend/src/main/java/com/ll/TeamProject/domain/user/entity/SiteUser.java
@@ -1,5 +1,6 @@
 package com.ll.TeamProject.domain.user.entity;
 
+import com.ll.TeamProject.domain.user.enums.Role;
 import com.ll.TeamProject.global.jpa.entity.BaseTime;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
