+++ backend/src/main/java/com/ll/TeamProject/domain/user/enums/Role.java
@@ -1,4 +1,4 @@
-package com.ll.TeamProject.domain.user.entity;
+package com.ll.TeamProject.domain.user.enums;
 
 public enum Role {
     ADMIN, // 관리자
